+++ llm_openai/models/__init__.py
@@ -1,2 +1,3 @@
 from . import openai_provider
 from . import mail_message
+from . import llm_model